refactor: consolidate environment and platform constants into a single module
