+++ src/main/windows/main.ts
@@ -2,7 +2,7 @@ import { BrowserWindow } from 'electron'
 import { join } from 'node:path'
 
 import { createWindow } from 'shared/lib/electron-app/factories/windows/create'
-import { ENVIRONMENT } from 'shared/constants/environment'
+import { ENVIRONMENT } from 'shared/constants'
 import { displayName } from '~/package.json'
 
 export async function MainWindow() {
