+++ src/shared/constants.ts
@@ -1,3 +1,7 @@
+export const ENVIRONMENT = {
+  IS_DEV: process.env.NODE_ENV === 'development',
+}
+
 export const PLATFORM = {
   IS_MAC: process.platform === 'darwin',
   IS_WINDOWS: process.platform === 'win32',
