--- src/shared/constants/environment.ts
@@ -1,3 +1,0 @@
-export const ENVIRONMENT = {
-  IS_DEV: process.env.NODE_ENV === 'development',
-}
