+++ src/shared/lib/electron-app/factories/app/setup.ts
@@ -6,8 +6,7 @@ import {
 } from 'electron-extension-installer'
 
 import { ignoreConsoleWarnings } from '../../utils/ignore-console-warnings'
-import { ENVIRONMENT } from 'shared/constants/environment'
-import { PLATFORM } from 'shared/constants/platform'
+import { PLATFORM, ENVIRONMENT } from 'shared/constants'
 import { makeAppId } from 'shared/utils'
 
 ignoreConsoleWarnings(['Manifest version 2 is deprecated'])
